requirements.txt
streamlit>=1.38.0
mysql-connector-python>=9.0.0
python-dotenv>=1.0.1
selenium>=4.24.0
